Expose more of the board's peripherals
Some use cases for the micro:bit require more of its peripherals to be
available. Examples include the 4tronix bit:bot, where the use of the
buzzer requires the use of p0 and a PWM instance. Using, e.g., the
wheels and the buzzer at the same time also requires multiple PWM
peripheral instances.